chore: score in seperate module
