--- tasks/catalyst/catalyst/tasks/adsorption/temp/adsorption_sites.json
@@ -1,69 +1,0 @@
-{
-  "ontop": [
-    [
-      3.8492813314670213,
-      2.2223791999573574,
-      13.393094518066786
-    ],
-    [
-      1.9246410522130029,
-      3.3335704333868716,
-      12.607363803125004
-    ]
-  ],
-  "bridge": [
-    [
-      2.8869611918400166,
-      2.777974816672116,
-      13.000229160595898
-    ],
-    [
-      2.886961950521591,
-      0.5555926331668026,
-      13.393094518066786
-    ],
-    [
-      1.9246418108945873,
-      1.1111882498815595,
-      13.000229160595895
-    ],
-    [
-      0.9623216712675772,
-      1.6667838665963146,
-      12.607363803124999
-    ]
-  ],
-  "hollow": [],
-  "all": [
-    [
-      3.8492813314670213,
-      2.2223791999573574,
-      13.393094518066786
-    ],
-    [
-      1.9246410522130029,
-      3.3335704333868716,
-      12.607363803125004
-    ],
-    [
-      2.8869611918400166,
-      2.777974816672116,
-      13.000229160595898
-    ],
-    [
-      2.886961950521591,
-      0.5555926331668026,
-      13.393094518066786
-    ],
-    [
-      1.9246418108945873,
-      1.1111882498815595,
-      13.000229160595895
-    ],
-    [
-      0.9623216712675772,
-      1.6667838665963146,
-      12.607363803124999
-    ]
-  ]
-}
