--- tasks/catalyst/catalyst/tasks/adsorption/temp/chosen_site.json
@@ -1,1 +1,0 @@
-[3.8492813314670213, 2.2223791999573574, 13.393094518066786]
