--- tasks/catalyst/catalyst/tasks/adsorption/temp/co2.cif
@@ -1,38 +1,0 @@
-# generated using pymatgen
-data_CO2
-_symmetry_space_group_name_H-M   'P 1'
-_cell_length_a   5.49707328
-_cell_length_b   5.49707328
-_cell_length_c   5.49707328
-_cell_angle_alpha   90.00000000
-_cell_angle_beta   90.00000000
-_cell_angle_gamma   90.00000000
-_symmetry_Int_Tables_number   1
-_chemical_formula_structural   CO2
-_chemical_formula_sum   'C4 O8'
-_cell_volume   166.10954147
-_cell_formula_units_Z   4
-loop_
- _symmetry_equiv_pos_site_id
- _symmetry_equiv_pos_as_xyz
-  1  'x, y, z'
-loop_
- _atom_site_type_symbol
- _atom_site_label
- _atom_site_symmetry_multiplicity
- _atom_site_fract_x
- _atom_site_fract_y
- _atom_site_fract_z
- _atom_site_occupancy
-  C  C0  1  -0.00000000  0.00000000  0.00000000  1
-  C  C1  1  0.50000000  0.00000000  0.50000000  1
-  C  C2  1  0.50000000  0.50000000  0.00000000  1
-  C  C3  1  -0.00000000  0.50000000  0.50000000  1
-  O  O4  1  0.12253698  0.12253698  0.12253698  1
-  O  O5  1  0.37746302  0.87746302  0.62253698  1
-  O  O6  1  0.62253698  0.37746302  0.87746302  1
-  O  O7  1  0.87746302  0.62253698  0.37746302  1
-  O  O8  1  0.87746302  0.87746302  0.87746302  1
-  O  O9  1  0.62253698  0.12253698  0.37746302  1
-  O  O10  1  0.37746302  0.62253698  0.12253698  1
-  O  O11  1  0.12253698  0.37746302  0.62253698  1
