--- tasks/catalyst/catalyst/tasks/adsorption/temp/slab.cif
@@ -1,34 +1,0 @@
-# generated using pymatgen
-data_Si
-_symmetry_space_group_name_H-M   'P 1'
-_cell_length_a   3.84927954
-_cell_length_b   3.84927840
-_cell_length_c   18.85753565
-_cell_angle_alpha   90.00000000
-_cell_angle_beta   90.00000000
-_cell_angle_gamma   60.00000753
-_symmetry_Int_Tables_number   1
-_chemical_formula_structural   Si
-_chemical_formula_sum   Si8
-_cell_volume   241.97716079
-_cell_formula_units_Z   8
-loop_
- _symmetry_equiv_pos_site_id
- _symmetry_equiv_pos_as_xyz
-  1  'x, y, z'
-loop_
- _atom_site_type_symbol
- _atom_site_label
- _atom_site_symmetry_multiplicity
- _atom_site_fract_x
- _atom_site_fract_y
- _atom_site_fract_z
- _atom_site_occupancy
-  Si  Si0  1  0.66666685  0.66666652  0.10416667  1.0
-  Si  Si1  1  0.00000011  0.99999991  0.06250000  1.0
-  Si  Si2  1  0.33333381  1.33333294  0.27083333  1.0
-  Si  Si3  1  -0.33333293  1.66666634  0.22916667  1.0
-  Si  Si4  1  0.00000077  1.99999937  0.43750000  1.0
-  Si  Si5  1  -0.66666597  2.33333276  0.39583333  1.0
-  Si  Si6  1  -0.33333226  2.66666580  0.60416667  1.0
-  Si  Si7  1  -0.99999900  2.99999919  0.56250000  1.0
